pyproject.toml
[project]
name = "ai-newsletter"
version = "0.1.0"
description = ""
authors = [
    {name = "BennyTaccardi",email = "benny.taccardi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic-ai (>=0.4.11,<0.5.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "openai (>=1.107.0,<2.0.0)",
    "pypdf (>=6.0.0,<7.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "perplexityai (>=0.12.0,<0.13.0)",
    "poetry-core (>=2.0.0)",
    "langchain-community (>=0.3.30,<0.4.0)",
    "google-genai (>=1.42.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "src", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
